refactor: split render functions into separate modules
